tickets module

+++ src/matches/matches.service.ts
@@ -1,13 +1,16 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, Inject, forwardRef, BadRequestException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Match } from './entities/match.entity';
 import { CreateMatchDto } from './dto/create-match.dto';
 import { Event } from 'src/events/entities/events.entity';
 import { TeamsService } from 'src/teams/teams.service';
-import { EventStatus, MatchStatus } from '../common/enums/common.enum';
+import { MatchStatus, ContestStatus } from '../common/enums/common.enum';
 import { Contest } from 'src/contests/entities/contest.entity';
 import { UpdateMatchDto } from './dto/update-match.dto';
+import { ContestsService } from 'src/contests/contests.service';
+
+@Injectable()
 export class MatchesService {
   constructor(
     @InjectRepository(Match)
@@ -17,6 +20,8 @@ export class MatchesService {
     @InjectRepository(Contest)
     private contestRepository: Repository<Contest>,
     private teamsService: TeamsService,
+    @Inject(forwardRef(() => ContestsService))
+    private contestsService: ContestsService,
   ) { }
 
   async createMatch(createMatchDto: CreateMatchDto): Promise<Match> {
@@ -91,33 +96,21 @@ export class MatchesService {
     }
 
     match.status = status;
+    if(status === MatchStatus.COMPLETED){
+      await this.contestsService.update(match.contests[0].id, {status: ContestStatus.COMPLETED});
+    }
     return this.matchRepository.save(match);
   }
 
 
   async deleteMatch(id: string): Promise<void> {
-    // if (updateEventDto.teamAId) {
-    //     const teamA = await this.teamsService.findOne(updateEventDto.teamAId);
-    //     if (!teamA) {
-    //       throw new NotFoundException(
-    //         `Team A with ID ${updateEventDto.teamAId} not found`,
-    //       );
-    //     }
-    //     event.teamA = { id: teamA.id } as any;
-    //     delete updateEventDto.teamAId;
-    //   }
-
-    //   // Handle teamBId update if provided
-    // //   if (updateEventDto.teamBId) {
-    // //     const teamB = await this.teamsService.findOne(updateEventDto.teamBId);
-    // //     if (!teamB) {
-    // //       throw new NotFoundException(
-    // //         `Team B with ID ${updateEventDto.teamBId} not found`,
-    // //       );
-    // //     }
-    // //     event.teamB = { id: teamB.id } as any;
-    // //     delete updateEventDto.teamBId;
-    // //   }
+    const match = await this.findOne(id);
+    if(!match){
+      throw new NotFoundException('Match not found');
+    }
+    if(match.status === MatchStatus.COMPLETED){
+      throw new BadRequestException('Cannot delete a completed match');
+    }
     await this.matchRepository.delete(id);
   }
 }
